Signup form, message board, styles, routing
Handled routing and CRUD operations for sign up form, adding messages,
and accessing the message board. If not logged in, you will only see
anonymous for the username on messages in the message board. You will
also be redirected from add-message (POST and GET) and flashed an error
message if not logged in. There are also other various flash messages
implemented.